+++ lib/src/client/query_response.dart
@@ -1,7 +1,23 @@
 import 'package:meta/meta.dart';
 import 'package:gql_exec/gql_exec.dart';
+import 'package:gql_link/gql_link.dart';
+import "package:collection/collection.dart";
 
 import './query_request.dart';
+
+enum DataSource {
+  /// A placeholder response source which can be used when waiting for another source
+  None,
+
+  /// Data originated from the client's [Link]
+  Link,
+
+  /// Data originated from the [Cache]
+  Cache,
+
+  /// Data originated from a user-provided [QueryRequest.optimisticResponse]
+  Optimistic,
+}
 
 /// Encapsulates a GraphQL query/mutation response, with typed
 /// input and responses, and errors.
@@ -10,33 +26,61 @@ class QueryResponse<T> {
   /// The event that resulted in this response
   final QueryRequest<T> queryRequest;
 
-  /// Whether or not the result is derived from an optimistic response
-  final bool optimistic;
+  /// The origin of the response.
+  final DataSource dataSource;
 
   /// The typed data of this response.
   final T data;
 
   /// The list of errors in this response.
-  final List<GraphQLError> errors;
+  final List<GraphQLError> graphqlErrors;
+
+  /// Any error returned by [Link]
+  final LinkException linkException;
 
   // TODO: is there a better way to implement loading?
-  bool get loading => errors == null && data == null;
+  bool get loading => linkException == null && data == null;
 
   /// If this response has any error.
-  bool get hasErrors => errors != null && errors.isNotEmpty;
+  bool get hasErrors =>
+      linkException != null &&
+      graphqlErrors != null &&
+      graphqlErrors.isNotEmpty;
 
   /// Instantiates a GraphQL response.
   const QueryResponse({
     @required this.queryRequest,
-    this.optimistic = false,
+    @required this.dataSource,
     this.data,
-    this.errors,
+    this.graphqlErrors,
+    this.linkException,
   });
+
+  List<Object> _getChildren() => [
+        queryRequest,
+        data,
+        graphqlErrors,
+        linkException,
+      ];
+
+  @override
+  bool operator ==(Object o) =>
+      o is QueryResponse &&
+      const DeepCollectionEquality().equals(
+        o._getChildren(),
+        _getChildren(),
+      );
+
+  @override
+  int get hashCode => const DeepCollectionEquality().hash(
+        _getChildren(),
+      );
 
   /// Creates a shallow copy
   QueryResponse.from(QueryResponse response)
       : queryRequest = response.queryRequest,
-        optimistic = response.optimistic,
+        dataSource = response.dataSource,
         data = response.data,
-        errors = response.errors;
+        graphqlErrors = response.graphqlErrors,
+        linkException = response.linkException;
 }