+++ CHANGELOG.md
@@ -7,6 +7,28 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.0] - 2020-04-22
+
+### Added
+
+- `LinkExceptions` are now caught and included in the `QueryResponse.linkException` property
+- added `QueryResponse.dataSource` property
+- added basic tests for GraphQL and network errors
+- added `Mutation` widget
+- added `executeOnListen` flag on `responseStream`
+
+### Changed
+
+- updated default subscription `FetchPolicy` to `CacheAndNetwork`
+- override equality for `QueryResponse`
+- update example
+
+### Removed
+
+- **BREAKING** removed optimistic flag on `QueryResponse` in favor of source property
+- **BREAKING** removed client error from `Query` widget
+- remove broken `FetchPolicy` tests
+
 ## [0.2.2] - 2020-02-29
 
 ### Changed